// src/app/Flex/DailyChecklist/models/PMS151_GetDailyChecklist_Detail_Item.ts
export class PMS151_GetDailyChecklist_Detail_Item {
    DAILY_CHECKLIST_HID: Number;
    MACHINE_NO: String;
    CHECKLISTITEMID: Number;
    CHECKLISTITEM_DESC: String;
    CHECK_FLAG: String;
    OK: Boolean;
    NG: Boolean;
    NG_REASON: String;
    REMARK: String;
    CHECK_REPH_ID: String;
    CREATEUSERID: String;
    CREATEMACHINE: String;
    CREATEDATETIME: Date;
    LASTUPDATEUSERID: String;
    LASTUPDATEMACHINE: String;
    LASTUPDATEDATETIME: Date;
}